[front] refactor: AvailableEntity moved in his own file and splitted into two component AvailaibleEntity and EntityNotAvailable
